Update to latest hackage-security
The hackage-security library now uses Int54 instead of Int everywhere (for
annoying JSON reasons). In this commit I prove a SafeCopy instance for Int54
which translates to Int, so that we remain backwards binary compatible. I do
change to Int54 in a few datatypes, but only in datatypes that are not (yet)
serialized on the running server.

+++ Distribution/Server/Packages/Metadata.hs
@@ -39,7 +39,7 @@ computePkgMetadata pkg revNo = (inIndexPkgMetadata pkgId, raw)
 
 pkgTarballTargets :: Int -> PackageIdentifier -> PkgTarball -> Sec.Targets
 pkgTarballTargets revNo pkgId pkgTarball = Sec.Targets {
-      targetsVersion     = Sec.FileVersion revNo
+      targetsVersion     = Sec.FileVersion (fromIntegral revNo)
     , targetsExpires     = Sec.expiresNever
     , targetsTargets     = Sec.FileMap.fromList [
                                (inRepoPkgTarGz pkgId, secFileInfo pkgTarballGz)
